# lib/api-zod/src/generated/types/addPhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Missionary Reporting API
 * OpenAPI spec version: 0.1.0
 */

export interface AddPhotoBody {
  url: string;
  /** @nullable */
  caption?: string | null;
}