# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Missionary Reporting API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from "./categoryCount";

export interface StatsResponse {
  totalReports: number;
  totalMissionaries: number;
  totalPeopleReached: number;
  totalLeadersTrained: number;
  totalCommunitiesServed: number;
  reportsByCategory: CategoryCount[];
}