# lib/api-zod/src/generated/types/timelineResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Missionary Reporting API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportWithDetails } from "./reportWithDetails";

export interface TimelineResponse {
  reports: ReportWithDetails[];
  total: number;
  hasMore: boolean;
}